Code Cleanup

+++ src/commands/RandomCommand.java
@@ -3,9 +3,9 @@ package commands;
 import controller.MasterController;
 
 public class RandomCommand extends OneInputCommand {
-    
+
     @Override
-    public double executeCommand() {
-        return (double)MasterController.myRandom.nextInt((int)myParameterOne);
+    public double executeCommand () {
+        return MasterController.myRandom.nextInt((int) myParameterOne);
     }
 }
